Try to circumvent MethodError
We're getting this MethodError:
```julia
MethodError(f=Base.Enums.namemap, args=(SciMLBase.ReturnCode.T,), world=0x0000000000009661)
```
Here I am trying to just print the `SciMLBase.ReturnCode.T` type so that
this is called once. However, it looks like it's not helping.

This commit should act as a reference for an issue though.